remove configuration
